Rename *-search-type => *-search-state
